requirements.txt
# Darkpool Monitor - Python Dependencies
# =============================================================================

# FastAPI & Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
httpx>=0.26.0

# HTTP Requests (sync version)
requests>=2.31.0

# Discord.py for bot commands
discord.py>=2.3.0

# Pydantic for data validation
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Optional: Data processing
pandas>=2.1.0
numpy>=1.26.0